requirements.txt
python-dotenv==1.0.1
colorlog==6.9.0
pytz==2025.1
socksio==1.0.0
zstandard==0.23.0
Brotli==1.1.0
h2==4.1.0
construct==2.10.70
httpx==0.28.1
Flask==3.1.0
asgiref==3.8.1
waitress==3.0.2